move band_center from utils to bandstructure.py
